moved PackageManager to System namespace
