+++ extensions/system/src/System/Package/Repository/ExtensionRepository.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Pagekit\Package\Repository;
+namespace Pagekit\System\Package\Repository;
 
+use Pagekit\Package\Repository\InstalledRepository;
 use Pagekit\Package\Loader\JsonLoader;
 use Pagekit\Package\Loader\LoaderInterface;
 
-class ThemeRepository extends InstalledRepository
+class ExtensionRepository extends InstalledRepository
 {
     /**
      * @var LoaderInterface
@@ -32,8 +33,10 @@ class ThemeRepository extends InstalledRepository
     {
         parent::initialize();
 
-        foreach (glob("{$this->path}/*/theme.json") as $config) {
-            $this->addPackage($this->loader->load($config));
+        if (empty($this->packages)) {
+            foreach (glob("{$this->path}/*/extension.json") as $config) {
+                $this->addPackage($this->loader->load($config));
+            }
         }
     }
 }
